move tempfile handling to caller of fetch/spider

+++ spiderfetch.py
@@ -2,29 +2,37 @@
 
 import os
 import sys
+import tempfile
 
 import fetch
 import filetype
 import spider
 import urlrewrite
 
+
 url = sys.argv[1]
 
+queue = [url]
+depth = 0
+while queue:
+    print "\nDepth: %s\nQueue: %s\n" % (depth, len(queue))
 
-filename = fetch.spider(url)
-data = open(filename, 'r').read()
-os.unlink(filename)
+    working_set = queue
+    queue = []
+    depth += 1
 
-urls = spider.unbox_it_to_ss(spider.findall(data))
-urls = urlrewrite.rewrite_urls(url, urls)
+    for url in working_set: 
+        try:
+            (_, filename) = tempfile.mkstemp(prefix=sys.argv[0] + ".")
+            fetch.spider(url, filename)
+            data = open(filename, 'r').read()
 
-for u in urls: 
-    try:
-        f = fetch.spider(u)
-        os.unlink(f)
-    except KeyboardInterrupt:
-        fetch.write_abort()
-        sys.exit(1)
-    except:
-        pass
-        #print "url failed:", u
+            urls = spider.unbox_it_to_ss(spider.findall(data))
+            urls = urlrewrite.rewrite_urls(url, urls)
+            queue.extend(urls)
+        except KeyboardInterrupt:
+            fetch.write_abort()
+            sys.exit(1)
+        finally:
+            if os.path.exists(filename):
+                os.unlink(filename)
